requirements.txt
numpy>=1.20.0
PyOpenGL>=3.1.5
PyOpenGL_accelerate>=3.1.5
glfw>=2.5.0
Pillow>=9.0.0
opencv-python>=4.5.0
imageio>=2.9.0